Separate out HIL mavlink interface (#582)
Separate out HIL mavlink interface into a separate class. 

Fix bug when default udp/tcp ports were not being set

Default TCP/UDP ports were not being set properly when there were no defaults in the sdf file
Use inline setter/getters for setting values
